ref: `QuestMadlibs.cs`
